Refactor; pull reading of the areas file aside.
Tt's persistance is different from the graph segments now; we don't want to load all areas for every request.
