+++ osm-jigsaw-api/app/graph/GraphReader.scala
@@ -3,51 +3,20 @@ package graph
 import java.io.BufferedInputStream
 import java.net.URL
 
-import model.{Area, GraphNode, OsmIdParsing, Point}
-import outputarea.OutputArea
+import javax.inject.Inject
+import model.{GraphNode, OsmIdParsing}
 import outputgraphnode.OutputGraphNode
 import play.api.Logger
 import progress.ProgressCounter
 
 import scala.collection.mutable
 
-class GraphReader extends OsmIdParsing {
+class GraphReader @Inject()(areasReader: AreasReader) extends OsmIdParsing {
 
   def loadGraph(areasFile: URL, graphFile: URL): GraphNode = {
 
-    def loadAreas(areasFile: URL): Map[Long, Area] = {
-
-      def outputAreaToArea(oa: OutputArea): Option[Area] = {
-        oa.id.flatMap { id =>
-          oa.area.map { a =>
-            val points = (oa.latitudes zip oa.longitudes).map(ll => Point(ll._1, ll._2)).toArray
-            Area(id = id, points = points, osmIds = oa.osmIds.map(toOsmId), area = a)
-          }
-        }
-      }
-
-      val areasMap = mutable.Map[Long, Area]()
-      val input = new BufferedInputStream(areasFile.openStream())
-      val counter = new ProgressCounter(step = 100000, label = Some("Reading areas"))
-      var ok = true
-      while (ok) {
-        counter.withProgress {
-          val added = OutputArea.parseDelimitedFrom(input).flatMap { oa =>
-            outputAreaToArea(oa).map { area =>
-              areasMap += area.id -> area
-              area
-            }
-          }
-          ok = added.nonEmpty
-        }
-      }
-      Logger.info("Mapped areas: " + areasMap.size)
-      input.close()
-      areasMap.toMap
-    }
-
     try {
-      val areas = loadAreas(areasFile)
+      val areas = areasReader.loadAreas(areasFile)
 
       def toGraphNode(ogn: OutputGraphNode): Option[GraphNode] = {
         ogn.area.flatMap { areaId =>
